Add a regression spec to cover sidekiq case
In 2.2.10 we introduced an issue where `bundle install` would unlock all
sources and their gems by default, and thus would endup re-resolving and
installing the latest version of gems, not the locked one.

This is somewhat inherent to the change of locking GEM sources
separately in the lockfile. If we change the lockfile format, the list
of sources changes, so we can't lock the previous specifications since
they come from an expired source.

I guess we could introduce a new mechanism to say "unlock the source of
this specification, but not its version". But it sounds like a
non-trivial feature with a very limited usage other than this migration.
So, the approach I took here is to keep the old format for existing
lockfiles and print a warning that it's not safe, so that people still
get `bundle install` working as before, but get a warning that they
should run `bundle update` or completely remove the lockfile and
regenerate it from scratch.

The above approach was implemented in previous commits. Here I'm just
adding specs to make sure it does what it's supposed to.